Rename module to be more descriptive.
